Move GetDeposits handler to Teku specific location
